Refactor Visitor to reflect new simplification
