+++ lib/msfl_visitors.rb
@@ -1,8 +1,8 @@
 require_relative 'msfl_visitors/ast'
-require_relative 'msfl_visitors/visitors/base'
 require_relative 'msfl_visitors/nodes'
+require_relative 'msfl_visitors/visitor'
 require_relative 'msfl_visitors/renderers/chewy_renderer'
-require_relative 'msfl_visitors/visitors/chewy_term_filter'
 require_relative 'msfl_visitors/parsers/msfl_parser'
+
 module MSFLVisitors
 end
