--- lib/msfl_visitors/visitors/chewy_term_filter.rb
@@ -1,7 +1,0 @@
-module MSFLVisitors
-  module Visitors
-    class ChewyTermFilter < Base
-
-    end
-  end
-end
